+++ lib/slather/project.rb
@@ -328,7 +328,13 @@ module Slather
         schemes_path = Xcodeproj::XCScheme.shared_data_dir(self.path)
         xcscheme_path = "#{schemes_path + self.scheme}.xcscheme"
 
-        raise StandardError, "No shared scheme named '#{self.scheme}' found in #{self.path}" unless File.exists? xcscheme_path
+        # Try to look inside 'xcuserdata' if the scheme is not found in 'xcshareddata'
+        if !File.file?(xcscheme_path)
+          schemes_path = Xcodeproj::XCScheme.user_data_dir(self.path)
+          xcscheme_path = "#{schemes_path + self.scheme}.xcscheme"
+        end
+
+        raise StandardError, "No scheme named '#{self.scheme}' found in #{self.path}" unless File.exists? xcscheme_path
 
         xcscheme = Xcodeproj::XCScheme.new(xcscheme_path)